+++ ui/src/components/collections/schema/SchemaFieldGeoPoint.svelte
@@ -1,0 +1,8 @@
+<script>
+    import SchemaField from "@/components/collections/schema/SchemaField.svelte";
+
+    export let field;
+    export let key = "";
+</script>
+
+<SchemaField bind:field {key} on:rename on:remove on:duplicate {...$$restProps} />