+++ ui/src/components/records/RecordUpsertPanel.svelte
@@ -23,6 +23,7 @@
     import SelectField from "@/components/records/fields/SelectField.svelte";
     import TextField from "@/components/records/fields/TextField.svelte";
     import UrlField from "@/components/records/fields/UrlField.svelte";
+    import GeoPointField from "@/components/records/fields/GeoPointField.svelte";
     import ImpersonatePopup from "@/components/records/ImpersonatePopup.svelte";
     import { confirm } from "@/stores/confirmation";
     import { setErrors } from "@/stores/errors";
@@ -730,6 +731,8 @@
                     <RelationField {field} {original} {record} bind:value={record[field.name]} />
                 {:else if field.type === "password"}
                     <PasswordField {field} {original} {record} bind:value={record[field.name]} />
+                {:else if field.type === "geoPoint"}
+                    <GeoPointField {field} {original} {record} bind:value={record[field.name]} />
                 {/if}
             {/each}
         </form>