+++ ui/src/components/records/fields/GeoPointValue.svelte
@@ -1,0 +1,9 @@
+<script>
+    export let value = {};
+</script>
+
+<div class="txt">
+    {value?.lon}
+    <span class="txt-disabled txt-xs">|</span>
+    {value.lat}
+</div>